Validate at least one prompt
