+++ src/components/Form/ValidatedFieldArray.js
@@ -1,0 +1,4 @@
+import { FieldArray } from 'redux-form';
+import withValidation from './withValidation';
+
+export default withValidation(FieldArray);
